requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0
requests==2.31.0
psycopg2-binary==2.9.9
dbt-postgres==1.7.4
great-expectations==0.18.8
pandas==2.1.4
